user login
Updated the User Login component to send the username and password to the back-end, which then checks the db to see if that information matches a user.  The back-end sends the login information and user_id back to the front-end.
The next step is to lift state to allow user information to live at the app-level.