Switched FK in StoryTopics in Model.py
I decided it made more sense to have the user_id as a FK in story_topics table, as opposed to the story_id FK set up at the beginning of the project.  This will allow me to save favorite topics to a specific user.

Renamed calls to server to start with /api/ for consistency
There are comments throughout files for updates from code review.