Logout and Conditionally Rendering Profile Page
Created logout function in server.py that successfully pops the user from the session.

Added ternary in profile page that tells shows a user's saved articles if logged in.  If user clicks "logout" while on the profile page, a message pops up telling user to sign in to see saved articles.